# openapi.yaml
openapi: 3.0.3
info:
  title: MineIntel AI REST API
  description: >
    Comprehensive REST API for MineIntel AI (Mining Intelligence, Regulatory Compliance
    & Automated Document Processing). Tailored for cross-platform Flutter clients targeting
    Android, iOS, and Windows Desktop.
  version: 1.0.0
  contact:
    name: MineIntel AI Engineering Team
servers:
  - url: https://mini-intel-ai-sih.vercel.app/api/v1
    description: Live Vercel Production REST API
  - url: http://10.0.2.2:5000/api/v1
    description: Android Emulator Local Bridge (Development)
  - url: http://127.0.0.1:5000/api/v1
    description: iOS Simulator & Windows Desktop Localhost (Development)
  - url: http://localhost:5000/api/v1
    description: Localhost Development Server
security:
  - BearerAuth: []
...
paths:
...
  /documents/upload:
    post:
      tags: [Documents]
      summary: Upload document (Multipart)
      description: Uploads document binary, computes SHA-256 hash for deduplication (returns 409 on duplicate), and queues background OCR/parsing.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file (PDF, DOCX, XLSX, CSV, PPTX, Image)
      responses:
        '201':
          description: Document uploaded and queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Document'
        '400':
          description: No file uploaded or invalid mime type
        '409':
          description: Duplicate document detected by checksum
  /documents:
    get:
      tags: [Documents]
      summary: List documents with filtering & pagination
      description: Non-admin users see only their own documents. Admins see all tenant documents.
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: type
          schema:
            type: string
            enum: [pdf, docx, xlsx, csv, image, pptx]
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, processing, completed, failed, extracted]
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: classification
          schema:
            type: string
            enum: [public, internal, confidential, restricted]
        - in: query
          name: dateFrom
          schema: { type: string, format: date-time }
        - in: query
          name: dateTo
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Paginated documents array
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDocumentResponse'
  /documents/{id}:
    get:
      tags: [Documents]
      summary: Get document details and extracted pages
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Document and extracted pages
    delete:
      tags: [Documents]
      summary: Delete document and associated chunks/files
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted document ID
  /documents/{id}/reviewers:
    put:
      tags: [Documents]
      summary: Replace reviewer assignments with optimistic concurrency (admin only)
      description: Read assignmentVersion from document details before saving. Legacy documents return zero. Old clients without a version are rejected; deploy updated clients with the server.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reviewerIds, assignmentVersion]
              properties:
                reviewerIds:
                  type: array
                  maxItems: 100
                  description: Active reviewer IDs; empty clears assignments. Case-insensitive duplicates are normalized.
                  items: { type: string, pattern: '^[a-fA-F0-9]{24}$' }
                assignmentVersion:
                  type: integer
                  minimum: 0
                  maximum: 9007199254740990
      responses:
        '200':
          description: Assignments replaced and version incremented atomically
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data:
                    type: object
                    required: [documentId, reviewerIds, assignmentVersion]
                    properties:
                      documentId: { type: string }
                      reviewerIds:
                        type: array
                        items: { type: string }
                      assignmentVersion: { type: integer, minimum: 1, maximum: 9007199254740991 }
        '400':
          description: INVALID_ID, INVALID_REVIEWERS, or INVALID_ASSIGNMENT_VERSION
        '401':
          description: Authentication required
        '403':
          description: Active admin required
        '404':
          description: DOCUMENT_NOT_FOUND
        '409':
          description: ASSIGNMENT_CONFLICT; no assignment or audit write. Reload and explicitly save again, never automatically retry.
  /documents/{id}/download:
    get:
      tags: [Documents]
      summary: Download original document file binary
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Binary file stream with Content-Disposition attachment
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /documents/{id}/metadata:
    get:
      tags: [Documents]
      summary: Retrieve technical & GIS metadata
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Metadata object
    put:
      tags: [Documents]
      summary: Update document metadata
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category: { type: string }
                classification:
                  type: string
                  enum: [public, internal, confidential, restricted]
                gisMetadata:
                  type: object
                  properties:
                    latitude: { type: number }
                    longitude: { type: number }
                    region: { type: string }
                retentionDate: { type: string, format: date-time }
      responses:
        '200':
          description: Updated document
  /documents/{id}/reprocess:
    post:
      tags: [Documents]
      summary: Trigger re-ingestion and OCR
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Queued for reprocessing
  /documents/{id}/status:
    get:
      tags: [Documents]
      summary: Ingestion progress & steps (Poll this endpoint)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ProcessingJob object
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProcessingJob'

  # =========================================================================
  # 9. EXTRACTION & HITL
  # =========================================================================
...
  /audit/document/{documentId}:
    get:
      tags: [Audit]
      summary: Audit trail for specific document
      parameters:
        - in: path
          name: documentId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Document audit logs
...
  /integration/documents:
    get:
      tags: [Integration]
      summary: DMS-compatible document catalog
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: fileType
          schema: { type: string }
      responses:
        '200':
          description: Paginated DMS documents
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Success
        data:
          type: object
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 48
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        pages:
          type: integer
          example: 1
...
    Document:
      type: object
      properties:
        _id: { type: string }
        filename: { type: string }
        originalName: { type: string }
        mimeType: { type: string }
        fileSize: { type: integer }
        fileType:
          type: string
          enum: [pdf, docx, xlsx, csv, image, pptx]
        hash: { type: string }
        category: { type: string }
        classification:
          type: string
          enum: [public, internal, confidential, restricted]
        status:
          type: string
          enum: [pending, processing, completed, failed, extracted]
        totalPages: { type: integer }
        gisMetadata:
          type: object
          properties:
            latitude: { type: number }
            longitude: { type: number }
            region: { type: string }
        retentionDate: { type: string, format: date-time }
        uploadedAt: { type: string, format: date-time }
        userId: { type: string }
        reviewerIds:
          type: array
          items: { type: string }
        assignmentVersion: { type: integer, minimum: 0, maximum: 9007199254740991, default: 0 }
    PaginatedDocumentResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ProcessingJob:
      type: object
      properties:
        _id: { type: string }
        documentId: { type: string }
        status:
          type: string
          enum: [queued, processing, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        currentStep: { type: string }
        steps:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              status:
                type: string
                enum: [pending, processing, completed, failed]
        startedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        error: { type: string }
...